backend/requirements.txt
Flask
flask-sqlalchemy
flask-migrate
flask-cors
openai
python-dotenv
Flask-Login
Flask-Bcrypt
psycopg2-binary
